feat: implement screenshot authentication and validation, including quota checks and response building for optimized screenshot API
